SymptoTrack/requirements.txt
openai-whisper
torch
numpy
scikit-learn
pandas
